Add "verify" folder.
The verify folder will be used for tests that are best performed by humans
instead of computers.

For example, error messages should be verified by humans. One of the
reason for this is that error messages are rarely (if ever) documented and
error messages have no guarantees on how they change in different times
and places. Humans can also notice things about error messages that are
difficult to automate such as the tone of the message, the appropriateness
of the message, or even unusual things like the use of a word that is fine
in one language but sounds like an inappropriate word in a different
language.

NOTE:
The contents of this folder is not guaranteed to stay the same.

The contents of this folder will change depending on the needs of people
verifying the correctness of the library functions.