Fix build script.
By default, build in a "Debug/" instead of the "Release/". This will allow
development to continue without affecting the contents of the "Release/"
directory.

The -Release switch must be used explicitly to build the "Release/"
directory.

Normalize line endings to crlf so that they can be read in notepad.
Hopefully, this will the scripts in the zip file (for download from
GitHub) contain crlf line endings.