Fix helpConverter.psm1.
Always check for null since properties of the help object may or may not
be present based on the content of the help file and also on the version
of PowerShell being used.

Properly handle the following:
    * multiple .Link directives
    * multiple .Inputs directives
    * multiple .Outputs directives.

Match the order of the xml elements with the order of xml elements in the
Microsoft PowerShell help files. This will, hopefully, make changes to the
xml files using different tools less noisy.