Change how the script version of the library is built.
The functions in the script version of the library will be wrapped inside
a dynamic module. This will remove the difficulties with invoking script
blocks that are passed to the functions.

See, for example, the initial version of Assert-All, where all variables
needed to be set to private and the effect of $ErrorActionPreference could
not be guaranteed.