Refactor assert functions.
Factor out the creation of messages and error records into their own
functions. These functions are private to the module and the names of
these functions are prepended with the module GUID. The reason is that if
someone really wants a script version of this library (a valid concern),
they can just remove the functions from the dynamic module and the names
of the private functions won't clash with other functions that they may
have in their own script or module.