Fix potential bug in assert functions.
Parameters that accept the value to be asserted must have a parameter type
of System.Object.

In PowerShell, parameter types don't mean:
    "make sure arguments passed in to this parameter are of this type"
but
    "if necessary, convert arguments passed in to this parameter to this
    type".

If the parameter type that accepts the value to be asserted has a type
other than System.Object, then you cannot be sure if the assert function
is testing the value or a new object that PowerShell created.

Note:
The implicit conversions are usually good for a lot of functions, but just
not for assert functions.

Important:
Creating a failing test for this type of error is not always possible.

The best that can be done is to create tests that prevent this type of
error. But I'm not even sure if that is always possible. Luckily, for this
case, it was possible.