Add logic functions.
Add logic functions so that the user of the library can write predicates
without having to deal with PowerShell's implicit conversions and
filtering semantics.

Many-Valued Logic
=================
These functions will return $true, $false, or $null. A return value of
$null means that the operation could not be performed.

Connectives
===========
The following logical connectives may be added in the future:
    * and
    * or
    * xor
    * nand
    * nor
    * xnor.

Negation will not be implemented as a single function. Instead, negation
will take many forms, such as NotTrue, NotFalse, NotNull, NotExists, and
so on. This gives us the chance to define and document the many meanings
of "not".

Error Conditions
================
If a logic function invokes a predicate, and that predicate throws an
error, the logic function must stop and propagate the error to the caller
of the logic function.

There are 3 reasons for doing this:
    * the error is evidence that the predicate does not work
    * the predicate may have state and that state may be corrupt
    * the predicate is the reponsibility of the function caller.

The ErrorAction parameter will have no effect on the logic functions.

The $ErrorActionPreference variable will only affect the arguments
supplied to the functions (the predicates) but the logic functions will
not be affected by the $ErrorActionPreference variable.