Fix parameters for comparison functions.
Remove the AllowEmptyCollection attribute for mandatory parameters of type
System.Object since that attribute does not seem to have any effect on
mandatory parameters of type System.Object.

Do not remove the AllowNull attribute. The AllowNull attribute does not
seem to have any effect on mandatory parameters of type System.Object, but
it has an effect on mandatory parameters of other types. For consistency
and documentation purposes, leave the attribute in the script. No need to
constantly guess which rule is in play when reading a script.