Comply with community style guidelines.
Put the closing brace in its own line.

Although not specified in the guidelines, do this only for the closing
brace of blocks that span multiple lines. Also, DO NOT do this for a group
of blocks, such as an array of script blocks passed to a function.

After all, changing this:
    Sort-Object -Property {...}, {...}
to this:
    Sort-Object -Property {
       ...
    }
    , {
       ...
    }
would be a little silly.