Update XML help files and generator.
Make sure that the syntax section always has at least one syntax item
element in order to comply with the schema.

Note:
Even though newer version of PowerShell always adds a syntax section to
the output of Get-Help, that was not always true in previous versions of
PowerShell and is not guaranteed to always be true in future versions of
PowerShell.

Source:
developerCommand.xsd in the PowerShell installation directory.