Update XML help files and generator.
Make sure that the inputs section always has at least one input type
element and always has a description element in order to comply with the
schema.

Remove blank optional elements in the input type section. The only
mandatory element for the input type is the uri.

Note:
Even though newer versions of PowerShell always adds an input section to
the output of Get-Help, that was not always true in previous versions of
PowerShell and is not guaranteed to always be true in future versions of
PowerShell.

Source:
developerCommand.xsd in the PowerShell installation directory.
developer.xsd in the PowerShell installation directory.