Update XML help files and generator.
Make sure that the outputs section always has at least one output type
element and always has a description element in order to comply with the
schema.

TODO: Add the optional "possibleValues" element.

Remove blank optional elements in the output type section. The only
mandatory element for the output type is the uri.

Note:
Even though newer versions of PowerShell always adds an output section to
the output of Get-Help, that was not always true in previous versions of
PowerShell and is not guaranteed to always be true in future versions of
PowerShell.

Source:
developerCommand.xsd in the PowerShell installation directory.
developer.xsd in the PowerShell installation directory.