Fix OutVariable bug in the comparison functions.
Do not use the "return" keyword to return items from the function because
the -OutVariable parameter in PowerShell version 2 will not work properly.

The "return" keyword can still be used to exit functions early. The
"return" keyword can also still be used normally in private functions
where the -OutVariable parameter will not be used.