Prepend external commands with its module name.
Prepend external commands with its module name so that if there are
multiple commands with the same name, the correct command will be called
by PowerShell.

Although it is unlikely that someone will create a function with a name
that will conflict with the name of a function that many people depend on
(such as New-Object, Where-Object, and Sort-Object) the likelyhood of such
an event happening has increased.