Extract getType function from getListElementType.
Add a getType function that will always call an object's "real" GetType
method since the properties and methods of an object can be overriden in
PowerShell. The mechanism that allows access to the "real" properties and
methods (the psbase property found in all objects) doesn't work for the
GetType method in PowerShell 4 and possibly PowerShell 3.

Also, have the getType function return System.Void for null. This is
probably controversial, but it's a lot more useful than a null reference
exception. Besides, in PowerShell, I usually don't worry about the types
of identifiers (only the types of objects) and specifying System.Void as
the "type" for null isn't really a problem, especially for a private
helper function for this module.