Add rotate functions to Group-ListItem.
Add -RotateLeft and -RotateRight parameters to Group-ListItem.

These functions are basically filtered forms of -Permute with no argument
for the -Size parameter.

It is very important that the rotate functions always output at least 1
group. See the note below.

Note
====
These functions should always output a group that is equivalent to the
input list. This is important for the case of an empty list.

For example, Group-ListItem -RotateLeft @(), can have the following
outcomes:
    * throw error
    * output nothing
    * output 1 group with 0 items.

Throw Error
-----------
Throwing an error for an empty list input is not good because an empty
list is still a list. It is a valid input.

Ouput Nothing
-------------
Outputing nothing will make modifying tests dangerous. For example, let's
say the user has this assertion:

    assert-true (invoke-somefunction $list)

and modifies it to this:

    foreach ($group in @(Group-ListItem -RotateLeft $list)) {
        assert-true (invoke-somefunction $group.Items)
    }

The idea of the modification is to perform the assertion for each rotation
of $list. If -RotateLeft outputs nothing for empty lists, then this
modification will skip the assertion instead of running the assertion once
with the empty list.

In other words, this is a ticking time bomb.

Output 1 group with no items
----------------------------
This does not have the problems mentioned above.
This is also consistent with the permute function.

commit 47babdc6d99bb91ac499fb99e0780a98420c7c9a
Author: Francis de la Cerna <imfrancisd@users.noreply.github.com>
Date:   Sat Aug 22 00:10:04 2015 -0500

    Update maintenance comment.

commit 17adbf845bff4cfbb82fc3d3aca05d14a25ad89b
Author: Francis de la Cerna <imfrancisd@users.noreply.github.com>
Date:   Sat Aug 22 00:03:47 2015 -0500

    Update XML help file.

commit 686a7844d7bcb587e0d51d996d2b226a823a8d3a
Author: Francis de la Cerna <imfrancisd@users.noreply.github.com>
Date:   Fri Aug 21 23:45:25 2015 -0500

    Make minor corrections to Group-ListItem help.

commit 233ecb67fd055f49b051fc01778ba4ddeddc5549
Author: Francis de la Cerna <imfrancisd@users.noreply.github.com>
Date:   Fri Aug 21 23:23:30 2015 -0500

    Add help to Group-ListItem -RotateLeft and -RotateRight.

commit 7e2c8044f372ddae6bc499a3083b0bc5215eb03d
Author: Francis de la Cerna <imfrancisd@users.noreply.github.com>
Date:   Sat Aug 15 21:22:05 2015 -0500

    Add Group-ListItem -RotateLeft and -RotateRight.