Set pipeline parameters with value in begin block.
Parameters that get their value from the pipeline are considered "not set"
if the pipeline is empty. If Strict Mode is on, then using that parameter
when it is "not set" will generate an error.

The simplest solution is just to set the parameter with a value before the
parameter takes pipeline input.

Note, however, that in PowerShell 2 (and maybe also in higher versions),
setting a pipeline parameter with a default value in the param block will
not work. If Strict Mode is on, the pipeline parameter must be set
implicitly by the process block, and/or explicitly in the begin/end blocks
before that parameter can be used.