Change StrictMode default to Off in build.ps1.
Since PowerShell *may* or *may not* set variables for you, and
since PowerShell *may* or *may not* add properties for you, and
since these things can change with different versions of PowerShell,
using StrictMode really requires extensive tests, both inside and outside
of the library.

    Functions that pass (with high confidence) StrictMode -version 2:
        Logic Functions
        Assert functions
        Most of the collection functions.

    Functions that *may* fail StrictMode -version 2:
        Comparison functions.

    Functions that pass StrictMode -Latest:
        I don't even know what StrictMode -Latest checks for.
        The only way to know is to test all functions for every possible
        thing imaginable.

However, not using Set-StrictMode at all, will cause the script to inherit
the ScriptMode setting from the parent scope.

So, for now, the default is Off.

The goal is to change it to StrictMode -Version 2.

See the discussion here:
    Strict Mode
    https://github.com/PoshCode/PowerShellPracticeAndStyle/issues/51