Add -Quantity parameter to *Exists help files.
Add the -Quantity parameter to the exists functions.

Examples:
    if (Test-Exists $numbers {param($n) $n -lt 0} -Quantity Any) {
        ...
    }

    Assert-Exists $numbers {param($n) $n -eq 5} -Quantity Single

    $numbers |
        Assert-PipelineExists    {param($n) $n -gt 1} -Quantity Single |
        Assert-PipelineNotExists {param($n) $n -eq 0} -Quantity Multiple |
	Out-Null

Uniqueness
==========
The -Quantity parameter adds the concept of uniqueness in the Exists
functions.

The words "Any", "Single", and "Multiple" were used to specify quantities
because "Any" and "Single" were already used in other function names.

-Quantity enum
==============
The values for -Quantity ("Any", "Single", "Multiple") cannot be combined.
For example, this is not allowed:

    #not allowed
    Assert-Exists $numbers {param($n) $n -eq 0} -Quantity Single, Multiple

because the meaning of that line is not obvious. Does it mean "single or
multiple", or does it mean "single and multiple", or does it mean "single
overrides multiple", or does it mean "multiple overrides single", or some
other interpretation.

If combining the values were allowed, the user will have to verify what
the function does over and over again because there are too many possible
interpretations.

-Quantity enum vs. -Count int
=============================
It was also possible to add a -Count parameter instead of -Quantity, but a
-Count parameter requires requires the additional concepts of upper and
lower limits so that the user can specify that the -Count is exact, is a
maximum limit, or a minimum limit (and also if those limits are inclusive
or exclusive). There are just too many things to add to the code and to
the tests if -Count is used.

Another reason for not adding -Count to the Exists functions is that I
would rather add a predicate to Assert-PipelineCount than add -Count to
the Exists functions. There are already too many flavors of exists (some,
many, few, more than, less than, exactly, has, contains, any, none, empty,
and so on), and all can be expressed with Assert-PipelineCount if
Assert-PipelineCount had a -Predicate parameter. So why bother with
Exists? Because it has a very widely known and a very widely accepted
meaning in other subjects outside of testing and programming.

-Quantity enum vs. -Unique switch
=================================
It was also possible to add a -Unique switch parameter instead of
-Quantity, but the problem was how to specify not unique. It was tempting
to do this:

    Assert-Exists $numbers {param($n) $n -eq 10} -Unique:$false

to specify "not unique", but using a switch parameter in this way is not
generally accepted in the PowerShell community. See the following
discussion on GitHub:

    Is there guidance for using the 3 states of a switch parameter?
    https://github.com/PoshCode/PowerShellPracticeAndStyle/issues/69

The biggest problem is that the assertions are difficult to read when
there are many negations involved:

    Assert-NotExists $numbers {param($n) $n -ne 10} -Unique:$false

which can be translated to something difficult to say. However, with
"Any", "Single", and "Multiple", the user has other options to write the
assertion in a less convoluted way.