Have quantifier functions pass 2 arguments to predicates.
The quantifier functions will pass 2 arguments to predicates:
* the first argument is the item to be examined
* the second argument is the index of the item to be examined.

So, the following will be possible:

    Assert-All $numbers {param($n, $index) $n -eq $expected[$index]}

Since script blocks can take any number of arguments (even if they don't
have any parameters), users can still supply script blocks that only have
one parameter.

Note:
=====
The problem is what to do if the index gets too large.

For the quantifier functions that expect a System.Collections.ICollection,
the index getting too large is not a problem because ICollection does not
have a Count property or method that can be larger than Int32. This
implicitly means that it is reasonable to expect that ICollection will not
have more than Int32.MaxValue elements.

The pipeline quantifier functions, however, are problematic because the
PowerShell pipeline does not have an implicit limit on the number of items
that can pass through it.

Of course, Int32.MaxValue is a very large number, and it is unlikely that
tests will contain collections or pipelines that contain Int32.MaxValue
elements.

On the other hand, these functions are made for tests, and tests have to
deal with unlikely events. For now, just have PowerShell throw an error in
the unlikely event that the index gets too large.

But how to handle this implicit limitation is still an open question.
Should this limitation be made explicit? Should it be documented? How
should it be tested?

Linq has a similar problem, but it seems like an index overflowing is left
undefined. See the Select method at:

    Enumerable.Select<TSource,TResult> Method(IEnumerable<TSource>,Func<TSource,Int32,TResult>)
    https://msdn.microsoft.com/en-us/library/bb534869(v=vs.110).aspx